consumer/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
pydantic==1.10.2
pymongo==4.2.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
pymongo[srv]==4.2.0
dnspython==2.0.0
configparser==5.3.0